Небольшие правки
- Альтернативная клавиша для предметов изменена на необязательный параметр
- Увеличен размер у песочков
- Тестовые сущности перемещены к тестам
- Небольшой рефакторинг
