--- Labirint.Core.Tests/Helpers/TestAbility.cs
@@ -1,15 +1,0 @@
-﻿namespace Labirint.Core.Tests.Helpers;
-
-internal class TestAbility(int? moveCount = null) : Ability
-{
-    public override string Name => "TestAbility";
-
-    public override int? MoveCount { get; } = moveCount;
-
-    public int HitCount { get; private set; }
-
-    public override void Hit()
-    {
-        HitCount++;
-    }
-}
